web: add keywords search on compositions
fix insidious parsing bug:
when parsing a template we build a parameter array
with items like x[name] = value.
If we find a parameter with no name,
we just append it with x[] = value,
which uses integer keys like x[0], x[1].

PHP 7 has the brain-damaged property that 0 == 'foobar'
(this is fixed in PHP 8, but come on!!).
So when we scan the keys of an argument array,
the integers match whatever we compare them with.
This leads to bad data.

Solution: instead of x[] = value, use x['positional_arg'] = value.
Such args are used only in attrib template calls.