Add basic support for expanding Mediawiki text:     expand_mw_text($str)
Expanding template calls: the set of templates (as an array of name => text)
is stored in a serialized file, templates.ser.
If expand_mw_text() encounters a template call,
if looks up the template and scans through it,
replacing parameters with arguments,
and handling pseudo-templates like #if: and #ifeq:.
This is all recursive, of course.

Links ([[pagename]] or [[pagename|text]]): parse them.
I'm not sure what to replace them with.

This is preliminary and incomplete.