web: finish adding ensemble to performance. Display and edit it in both contexts: concert editing and recording editing.
Note: a performance can have at most one ensemble.
That's OK 99.99% of the time.

In concert editing, put ensemble in a separate column.

When showing performances of a composition,
choose which columns to display based on the performances.

Remove performance.is_recording.
If there are files, it's a recording.

Flesh out the ensemble item page.
TODO: make it editable.

Technical note: MySQL treats the string 'null' as valid JSON
(but not, e.g., the empty string).  Go figure.